Meshtastic/Raspberry_RECEIVER/requirements.txt
# ============================================================================
# Requirements per OriBruni Raspberry RECEIVER
# ============================================================================
# Installazione: pip install -r requirements.txt
# 
# Per Raspberry Pi, potrebbe essere necessario installare prima alcune
# dipendenze di sistema:
# sudo apt update
# sudo apt install python3-pip python3-dev build-essential
# sudo apt install libmariadb-dev pkg-config
# sudo apt install i2c-tools python3-smbus
# ============================================================================

# Framework Web
Flask==2.3.3
Flask-SocketIO==5.3.6

# Database MySQL/MariaDB
mysql-connector-python==8.1.0

# Comunicazione Meshtastic
meshtastic==2.2.15

# Display LCD I2C
RPLCD==1.3.1

# Gestione interfacce di rete
netifaces==0.11.0

# Librerie di supporto per WebSocket
python-socketio==5.8.0
python-engineio==4.7.1

# Gestione configurazione
configparser==6.0.0

# Logging avanzato
colorlog==6.7.0

# Gestione date e tempo
python-dateutil==2.8.2

# Librerie per GPIO e I2C (Raspberry Pi)
RPi.GPIO==0.7.1
smbus2==0.4.2

# Librerie per monitoraggio sistema
psutil==5.9.5

# Gestione JSON avanzata
jsonschema==4.19.1

# Librerie per networking
requests==2.31.0
urllib3==2.0.4

# Gestione threading e concorrenza
eventlet==0.33.3

# Librerie per sviluppo e debugging
python-dotenv==1.0.0

# Sicurezza e crittografia
cryptography==41.0.4
bcrypt==4.0.1

# Gestione file e path
pathlib2==2.3.7

# Librerie per testing (opzionali)
pytest==7.4.2
pytest-flask==1.2.0

# Librerie per backup e compressione
zipfile36==0.1.3

# Gestione processi
supervisor==4.2.5

# ============================================================================
# Note per l'installazione su Raspberry Pi:
# ============================================================================
# 
# 1. Aggiornare il sistema:
#    sudo apt update && sudo apt upgrade -y
#
# 2. Installare dipendenze di sistema:
#    sudo apt install python3-pip python3-dev python3-venv
#    sudo apt install build-essential libmariadb-dev pkg-config
#    sudo apt install i2c-tools python3-smbus
#    sudo apt install git curl wget
#
# 3. Abilitare I2C:
#    sudo raspi-config
#    # Interfacing Options -> I2C -> Enable
#    # Oppure: echo 'dtparam=i2c_arm=on' | sudo tee -a /boot/config.txt
#
# 4. Creare ambiente virtuale:
#    python3 -m venv venv
#    source venv/bin/activate
#
# 5. Installare requirements:
#    pip install --upgrade pip
#    pip install -r requirements.txt
#
# 6. Per MySQL/MariaDB:
#    sudo apt install mariadb-server mariadb-client
#    sudo mysql_secure_installation
#
# 7. Configurare database:
#    sudo mysql -u root -p
#    CREATE DATABASE OriBruniRadioControls;
#    CREATE USER 'meshdash'@'localhost' IDENTIFIED BY 'your_password';
#    GRANT ALL PRIVILEGES ON OriBruniRadioControls.* TO 'meshdash'@'localhost';
#    FLUSH PRIVILEGES;
#
# ============================================================================
